Move tests for table/function name extraction to `test_db.py`
The code they're testing moved
